==> 4_predict.ipynb <==
from keras.models import load_model
import pandas as pd
import numpy as np
from extras import *

model = load_model('model_toni.h5')

df = pd.read_pickle('labeled_data.pkl')

#input_column = 'inception_v3_input'
width = 224
height = 224
input_column = 'vgg19_input'
add_normalized_column(df, width, height, 'original', input_column)
xs = df[input_column].as_matrix()
xs = np.array([ e for e in xs])
n_samples = len(xs)

preds = model.predict(xs, batch_size=25)

predicted_labels = np.argmax(preds, 1)
df['toni_ypred'] = None
df['toni_ypred'] = predicted_labels

# df.to_pickle('predicted_data.pkl')

df.columns
